+++ src/api/responses.ts
@@ -8,40 +8,54 @@ const DEFAULT_HEADERS = {
 const metrics = new Metrics('API Gateway');
 
 export function success(payload?: any, replacer?: (this: any, key: string, value: any) => any): ApiResponse {
-  const response = { statusCode: 200, headers: DEFAULT_HEADERS };
+  const response = { statusCode: 200, headers: { ...DEFAULT_HEADERS } };
   if (payload) {
     response['body'] = replacer ? JSON.stringify(payload, replacer) : JSON.stringify(payload);
+    response.headers['Content-Type'] = 'application/json';
   }
   metrics.success(response);
   return response;
 }
 
 export function invalid(errors?: any[]): ApiResponse {
-  const response = { statusCode: 400, headers: DEFAULT_HEADERS };
+  const response = { statusCode: 400, headers: { ...DEFAULT_HEADERS } };
   if (errors) {
     response['body'] = JSON.stringify({ errors });
+    response.headers['Content-Type'] = 'application/json';
   }
   metrics.invalid(response);
   return response;
 }
 
 export function notFound(message?: string): ApiResponse {
-  const response = { statusCode: 404, headers: DEFAULT_HEADERS };
+  const response = { statusCode: 404, headers: { ...DEFAULT_HEADERS } };
   if (message) {
     response['body'] = JSON.stringify({ message });
+    response.headers['Content-Type'] = 'application/json';
+  }
+  metrics.invalid(message);
+  return response;
+}
+
+export function notAuthorized(message?: string): ApiResponse {
+  const response = { statusCode: 403, headers: { ...DEFAULT_HEADERS } };
+  if (message) {
+    response['body'] = JSON.stringify({ message });
+    response.headers['Content-Type'] = 'application/json';
   }
   metrics.invalid(message);
   return response;
 }
 
 export function error(message?: any): ApiResponse {
-  const response = { statusCode: 500, headers: DEFAULT_HEADERS };
+  const response = { statusCode: 500, headers: { ...DEFAULT_HEADERS } };
   if (message && message instanceof Error) {
     logger.debug('Encountered error while processing request', message);
     message = message.message;
   }
   if (message) {
     response['body'] = JSON.stringify({ message });
+    response.headers['Content-Type'] = 'application/json';
   }
   metrics.error(response);
   return response;