+++ CHANGELOG.md
@@ -4,11 +4,25 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 
+## [3.1.2]&nbsp;&nbsp;(2020-02-11)
+
+### Fixed
+
+- Support additional properties on content-type
+- Add correct content-type header for JSON stringified API responses
+- Update old libraries
+
+## [3.1.1]&nbsp;&nbsp;(2020-01-22)
+
+### Added
+
+- Add a 'not authorized' response handler to the api wrapper
+
 ## [3.1.0]&nbsp;&nbsp;(2020-01-20)
 
 ### Changed
 
-- Missing signature params as undefined instead of null to allow for default values
+- Default signature params to undefined instead of null to allow for default values
 - Updated libraries
 
 ## [3.0.0]&nbsp;&nbsp;(2020-01-05)
@@ -156,7 +170,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/), 
 - Update older libraries
 - Now publish from Git tags instead of master pushes
 
-[3.0.1]: https://github.com/manwaring/lambda-wrapper/compare/v3.0.0...v3.1.0
+[3.1.2]: https://github.com/manwaring/lambda-wrapper/compare/v3.1.1...v3.1.2
+[3.1.1]: https://github.com/manwaring/lambda-wrapper/compare/v3.1.0...v3.1.1
+[3.1.0]: https://github.com/manwaring/lambda-wrapper/compare/v3.0.0...v3.1.0
 [3.0.0]: https://github.com/manwaring/lambda-wrapper/compare/v2.1.1...v3.0.0
 [2.1.1]: https://github.com/manwaring/lambda-wrapper/compare/v2.1.0...v2.1.1
 [2.1.0]: https://github.com/manwaring/lambda-wrapper/compare/v2.0.1...v2.1.0